+++ src/libstd/owned.rs
